refactor: EclipseCipher#deserialize EclipseCipher#serialize to int

+++ src/main/java/io/github/spigotrce/eclipsecipher/EclipseCipher.java
@@ -1,6 +1,7 @@
 package io.github.spigotrce.eclipsecipher;
 
 import java.security.SecureRandom;
+import java.util.Random;
 
 public class EclipseCipher {
     public static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/ ";
@@ -17,17 +18,17 @@ public class EclipseCipher {
         return stringBuilder.toString();
     }
 
-    public static String deserialize(long i) {
+    public static String deserialize(int i) {
         StringBuilder r = new StringBuilder();
         while (i > 0) {
-            r.append(CHARACTERS.charAt((int) (i % CHARACTERS.length())));
+            r.append(CHARACTERS.charAt(i % CHARACTERS.length()));
             i /= CHARACTERS.length();
         }
         return r.reverse().toString();
     }
 
-    public static long serialize(String t) {
-        long r = 0;
+    public static int serialize(String t) {
+        int r = 0;
         for (char c : t.toCharArray())
             r = r * CHARACTERS.length() + CHARACTERS.indexOf(c);
         return r;
